Convert remaining `BuildApk` into `CollectNativeFilesForArchive`.
